requirements.txt
numpy==1.23.5
pandas==1.5.3
requests==2.28.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==1.4.39
Flask>=2.0.0
Werkzeug>=2.0.0
free-proxy==1.1.1